Add bounded thread pool
