+++ packages/sync/src/lib.rs
@@ -1,3 +1,4 @@
 mod thread_pool;
 
-pub use self::thread_pool::ThreadPool;
+pub use self::thread_pool::bounded::BoundedThreadPool;
+pub use self::thread_pool::infinite::ThreadPool;
