+++ packages/sync/src/thread_pool/types.rs
@@ -6,8 +6,14 @@ pub enum Message {
     Terminate,
 }
 
+#[derive(Debug)]
+pub enum ExecuteError {
+    Full,
+}
+
 pub type Job = Box<dyn FnBox + Send + 'static>;
 
 pub type MessageSender = mpsc::Sender<Message>;
+pub type MessageSyncSender = mpsc::SyncSender<Message>;
 pub type MessageReceiver = mpsc::Receiver<Message>;
 pub type SharedMessageReceiver = Arc<Mutex<MessageReceiver>>;
